# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BigBack API
  version: 0.0.1
  description: REST API for the BigBack taste blending app (skeleton).
  license:
    name: MIT
    url: https://opensource.org/license/mit/
servers:
  - url: http://localhost:3000
security: []
paths:
  /auth/signup:
    post:
      summary: Create account
      operationId: authSignup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 32
                password:
                  type: string
                  minLength: 8
                  maxLength: 200
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
        "409":
          description: Username taken
        "500":
          description: Server error
  /auth/login:
    post:
      summary: Login
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 32
                password:
                  type: string
                  minLength: 8
                  maxLength: 200
      responses:
        "200":
          description: OK
        "400":
          description: Validation error
        "401":
          description: Invalid credentials
        "500":
          description: Server error
  /health:
    get:
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  service:
                    type: string
                  time:
                    type: string
              example:
                ok: true
                service: maillardmap-api
                time: 2026-03-31T00:00:00.000Z
        "400":
          description: Bad request
        "500":
          description: Server error
  /support/contact:
    post:
      summary: Public support form (marketing site; email via Resend)
      operationId: supportContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, subject, message]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                email:
                  type: string
                  format: email
                  maxLength: 254
                subject:
                  type: string
                  minLength: 1
                  maxLength: 200
                message:
                  type: string
                  minLength: 1
                  maxLength: 8000
                website:
                  type: string
                  description: Honeypot; if non-empty, server returns 204
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "204":
          description: Honeypot triggered
        "400":
          description: Validation error
        "429":
          description: Rate limited
        "503":
          description: Email not configured
        "500":
          description: Server error